long -> int
This reverts commit d0b93f34ef17480a35ec51644b486606aca49852.

+++ snackbar/Customer.java
@@ -2,8 +2,8 @@ package snackbar;
 
 public class Customer {
 	// fields
-	private static long maxId = 0;
-	private long id;
+	private static int maxId = 0;
+	private int id;
 	private String name;
 	private double cash;
 
@@ -15,7 +15,7 @@ public class Customer {
 		this.cash = cash;
 	}
 
-	public long getId() {
+	public int getId() {
 		return id;
 	}
 
@@ -35,7 +35,7 @@ public class Customer {
 		this.cash = cash + number;
 	}
 
-	public void buySnacks(Snack snack, long number) {
+	public void buySnacks(Snack snack, int number) {
     // remove snacks from machine
     snack.sellSnacks(number);
     this.cash -= number * snack.getCost();
